+++ src/console/controllers/MigrateController.php
@@ -20,7 +20,7 @@ class MigrateController extends \yii\console\controllers\MigrateController
         'davidhirtz\yii2\skeleton\migrations',
     ];
 
-    public string $dbFile = '@config/db.php';
+    public string $dbFile = '@root/config/db.php';
     public $templateFile = '@skeleton/views/migration.php';
 
     private ?array $_dbConfig = null;