+++ src/console/controllers/ParamsController.php
@@ -15,7 +15,7 @@ class ParamsController extends Controller
 {
     use ConfigTrait;
 
-    public string $config = '@config/params.php';
+    public string $config = '@root/config/params.php';
 
     /**
      * Setup application.